Revert "Pinjan versio"
