--- Assets/Prefabs/Movable Objects/Devices.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2cdd565f786341c45beaa7b513fd8ac7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
