--- Assets/Prefabs/Movable Objects/Devices/ChargeableDevice 1.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: df65cfdfdb7629943865fae035cc32cf
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
