--- Assets/Prefabs/Movable Objects/Devices/ChargeableDevice.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d72354561941b1e47a2484007c74e04e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
