--- Assets/Prefabs/Movable Objects/Devices/DeviceDoor.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4e57be145a8034f43bfa5568b59a0bdc
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
