--- Assets/Prefabs/Movable Objects/Devices/Separate parts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 06f233175fdade84b9b7703026c0ddaa
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
