--- Assets/Prefabs/Movable Objects/Devices/Separate parts/DeviceDoor.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 610d2757f0dc1fa45b80e24c9fbd3ee4
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
