--- Assets/Prefabs/Movable Objects/Devices/Separate parts/DeviceTarget.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 7704106d675a4c84687864d98f71a95b
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
