--- Assets/Prefabs/Movable Objects/Devices/Separate parts/DoorExitTrigger.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 594f177a0cfafbb4faf08c347c07107a
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
