--- Assets/Prefabs/Movable Objects/Devices/Separate parts/DoorTrigger.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 2769b55c2d4d88343afc5fdcf4a271de
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
