--- Assets/Prefabs/Movable Objects/Devices/Separate parts/Waypoint.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 7a8e7460510729f4284a532951d59239
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
