--- Assets/Script/PlayerAbilities/Kipina/Device.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 39499e81e96d6d54c8c36ce8d51c74aa
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
