--- Assets/Script/PlayerAbilities/Kipina/Device/DeviceOpenDoor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f50b5ad3e5ace1d43af83276fab0b872
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
