--- Assets/Script/PlayerAbilities/Kipina/Device/DeviceTargetScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 28641f02c449bc14bb0e1ea193ed521b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
