--- Assets/Script/PlayerAbilities/Kipina/Device/DoorExitTriggerScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a2f4af77022610240966af123763a8a8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
