--- Assets/Script/PlayerAbilities/Kipina/Device/DoorTriggerScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2bf727bddc7b9ce4483b052c8dcd6b8e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
