--- Assets/Script/PlayerAbilities/Kipina/PlayerCarryCable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 191f73ee986e17746a07668adab6302f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
