+++ Assets/Script/PlayerControl.inputactions
@@ -67,15 +67,6 @@
                     "processors": "",
                     "interactions": "",
                     "initialStateCheck": false
-                },
-                {
-                    "name": "CarryCable",
-                    "type": "Button",
-                    "id": "2181f636-2635-42f6-90de-3e72cd520a0c",
-                    "expectedControlType": "Button",
-                    "processors": "",
-                    "interactions": "",
-                    "initialStateCheck": false
                 }
             ],
             "bindings": [
@@ -395,39 +386,6 @@
                     "processors": "",
                     "groups": "Player2",
                     "action": "Charge",
-                    "isComposite": false,
-                    "isPartOfComposite": false
-                },
-                {
-                    "name": "",
-                    "id": "4898c3b2-5c4b-4623-bf1d-744a91996e58",
-                    "path": "<Gamepad>/buttonEast",
-                    "interactions": "",
-                    "processors": "",
-                    "groups": "Player1;Player2",
-                    "action": "CarryCable",
-                    "isComposite": false,
-                    "isPartOfComposite": false
-                },
-                {
-                    "name": "",
-                    "id": "840121b5-e2ca-4f46-bb52-63bd5b414fb5",
-                    "path": "<Keyboard>/h",
-                    "interactions": "",
-                    "processors": "",
-                    "groups": "Player1",
-                    "action": "CarryCable",
-                    "isComposite": false,
-                    "isPartOfComposite": false
-                },
-                {
-                    "name": "",
-                    "id": "b0c06038-7794-4567-b103-f5aefb342fa6",
-                    "path": "<Keyboard>/l",
-                    "interactions": "",
-                    "processors": "",
-                    "groups": "Player2",
-                    "action": "CarryCable",
                     "isComposite": false,
                     "isPartOfComposite": false
                 }
